Move most expressions into separate files
